refactor(server): extract pickAccept into internal accept module
