+++ packages/server/src/page-action-handler.ts
@@ -25,6 +25,7 @@ import {
   isAsyncGenerator,
 } from './sse.js';
 import type { ActionEntry } from './page-action-resolvers.js';
+import { pickAccept, type Accept } from './accept.js';
 import type { VNode } from 'preact';
 
 export interface PageActionHandlerOptions {
@@ -74,44 +75,6 @@ export interface PageActionHandlerOptions {
    * 500; this is a side channel only.
    */
   onError?: (err: unknown, ctx: { module: string; action: string }) => void;
-}
-
-type Accept = 'html' | 'json' | 'event-stream';
-
-/**
- * Parse the Accept header and return the highest-priority recognized type.
- * Follows RFC 9110 quality values (;q=0.9 etc) so that
- * "application/json, text/event-stream;q=0.9" correctly resolves to 'json',
- * not 'event-stream'. Unspecified quality defaults to 1.0.
- *
- * Internal: exported for unit testing only; not part of the public API.
- */
-export function pickAccept(header: string | undefined): Accept {
-  const h = header ?? '';
-  type Candidate = { type: Accept; q: number };
-  const candidates: Candidate[] = [];
-
-  for (const part of h.split(',')) {
-    const [mediaType, ...params] = part.trim().split(';');
-    const mt = (mediaType ?? '').trim().toLowerCase();
-    let q = 1.0;
-    for (const p of params) {
-      const kv = p.trim().split('=');
-      if (kv[0]?.trim() === 'q' && kv[1] !== undefined) {
-        const parsed = Number(kv[1].trim());
-        if (!Number.isNaN(parsed)) q = parsed;
-      }
-    }
-    if (mt === 'text/event-stream')
-      candidates.push({ type: 'event-stream', q });
-    else if (mt === 'application/json') candidates.push({ type: 'json', q });
-    else if (mt === 'text/html' || mt === '*/*')
-      candidates.push({ type: 'html', q });
-  }
-
-  if (candidates.length === 0) return 'html';
-  candidates.sort((a, b) => b.q - a.q);
-  return candidates[0]!.type;
 }
 
 async function parseBody(
